feat(auth-pages): added signin and signup pages (#4)
* feat(auth-pages): added signin and signup pages

This commit adds the SignIn and SignUp pages to the app, along with the necessary styling and UI components. The SignIn page includes a login form with email and password fields, as well as a "Forgot password" link. The SignUp page includes a registration form with fields for name, username, email, password, and birthday. Both pages use a consistent color scheme and typography to maintain a cohesive look and feel with the rest of the app. This feature will provide users with a streamlined and user-friendly way to sign in and sign up for the app.

* chore(root): removed unnecessary library

This commit removes the unnecessary library, date-fns, that is not being used.

* chore(auth-pages): added a constant file

This commit adds a constants file called authTexts.js and refactors the AuthSection component to use the text objects defined in the constants file. The authTexts.js file defines two objects, signUp and signIn, that contain the title, description, and button text for the sign up and sign in screens, respectively. These objects are now imported and used in the AuthSection component, making it easier to update and maintain the text in the future. This change improves the readability and maintainability of the codebase.

* chore(auth-pages): removed app.scss file

This commit removes the App.scss file and consolidates all of the CSS into the index.scss file. The CSS rules from App.scss have been moved into index.scss, which is the main CSS file for the application.

* feat(auth-pages): replace date picker with input type "date"

This commit removes the MUI material date picker component and replaces it with the native HTML5 input element with the "date" type. The dependencies for the MUI material date picker have also been uninstalled.