chore(auth-pages): changed function definition styles
Firstly, this commit changes the function definition to use the const = () => style, which is a more modern and concise way of defining functions. Additionally, it removes the unnecessary "dotenv" package and updates the importing statements to use local imports after global imports.

Moreover, the commit replaces the "a" element with the Link component.